+++ API/src/categories/model/category.rs
@@ -6,5 +6,5 @@ pub struct Category {
     pub category: String,
     pub type_: String,
     pub currency: Option<String>,
-    pub amount_cents: Option<i64>,
+    pub amount_cents: Option<i32>,
 }
